USERPOSTCREATE is compeletede

+++ src/Profex-ViewModels/Categories/CategoryViewModel.cs
@@ -2,10 +2,9 @@
 
 public class CategoryViewModel
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; } = String.Empty;
     public string Description { get; set; } = String.Empty;
-
     public  DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set;}
